+++ OpenACC/sample_code/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2026-6-15
+### Added
+* Add the contributions of the boundary condition to `10_jacobi/src/c`, `10_jacobi/src/fortran`, `11_jacobi-nvtx/src/c`, and `11_jacobi-nvtx/src/fortran`.
+
+### Fixed
+* Correct the script file to execute the program in `11_jacobi-nvtx/tests/fortran/acc2`.
 
 ## [1.0.0] - 2026-3-7
 ### Changed